pyproject.toml
[tool.poetry]
name = "mlops"
version = "0.1.0"
description = ""
authors = ["zagorulko <zagorulkooa@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
razdel = "^0.5.0"
transformers = "^4.28.1"
torch = "^2.0.0"
ftfy = "^6.1.1"
pypdf2 = "^3.0.1"
arxiv = "^1.4.7"
dvc-s3 = "^2.22.0"
sklearn = "^0.0.post4"
scikit-learn = "0.24.2"
click = "^8.1.3"
black = "^23.3.0"
flake8 = "^6.0.0"
mypy = "^1.3.0"
isort = "^5.12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
